requirements.txt
numpy>=1.18.0
python-dotenv>=0.20.0
openai>=0.27.0
pydantic>=1.9.0
requests>=2.26.0
httpx>=0.23.0 